clean up pdf interfaces
