+++ src/Service/Report/Document/Pdf/Layout/TableRowLayout.php
@@ -12,8 +12,8 @@
 namespace App\Service\Report\Document\Pdf\Layout;
 
 use App\Service\Report\Document\Layout\TableRowLayoutInterface;
-use App\Service\Report\Document\Pdf\Interfaces\PdfDocumentInterface;
 use App\Service\Report\Document\Pdf\Layout\Base\StatefulColumnedLayout;
+use App\Service\Report\Document\Pdf\PdfDocumentInterface;
 
 class TableRowLayout extends StatefulColumnedLayout implements TableRowLayoutInterface
 {
