+++ src/Service/Report/IssueReport/Interfaces/PrintFactoryInterface.php
@@ -12,7 +12,7 @@
 namespace App\Service\Report\IssueReport\Interfaces;
 
 use App\Service\Report\Document\Layout\Base\PrintableLayoutInterface;
-use App\Service\Report\Document\Pdf\Interfaces\PdfPageLayoutInterface;
+use App\Service\Report\Document\Pdf\PdfPageLayoutInterface;
 use App\Service\Report\Document\Transaction\Base\DrawableTransactionInterface;
 use App\Service\Report\IssueReport\Model\MetaData;
